Improved hasPermission readability
